#38 rename screens
